Postponed adding a test for removeIconsOnChrome,
until browser.extension.getURL gets implemented in jest-webextension-mock
