+++ source/lib/browser_functions.ts
@@ -1,4 +1,4 @@
-import { Tabs, browser } from 'webextension-polyfill-ts'
+import { Menus, Tabs, browser } from 'webextension-polyfill-ts'
 
 export const openInIncognitoWindow = (url: string, _tab: Tabs.Tab | undefined): void => {
   browser.windows.create({ incognito: true, url }).catch((error: Error) => {
@@ -21,6 +21,20 @@ export const blockZoomClientDownload = (blockClientDownload: boolean): void => {
       ['blocking'],
     )
   }
+}
+
+/**
+ * Solution suggested by glazou
+ * http://www.glazman.org/weblog/dotclear/index.php?post/2018/06/07/Browser-detection-inside-a-WebExtension
+ */
+export const removeIconsOnChrome = (
+  contextMenueProps: Menus.CreateCreatePropertiesType,
+): Menus.CreateCreatePropertiesType => {
+  const extensionUrl = browser.extension.getURL('/')
+  if (extensionUrl.startsWith('chrome')) {
+    delete contextMenueProps.icons
+  }
+  return contextMenueProps
 }
 
 export const sendErrorToDialog = (message: string, tab: Tabs.Tab | undefined): void => {
